apis: validate sheet names according to the Excel XLSX format
Sheet names must be UTF-8 encoded, with a length in the range [1, 31].
They should not start or end with "'", and cannot contain any of "*",
"/", ":", "?", "[", "\", and "`".

This change validates sheet names according to these rules. The `Sheets`
method of the `connectors` package now ensure that no invalid field
names are returned.