apis/connectors: updated validation for identity and timestamp column
This commit follows up on commit d63d01c. It enforces stricter
validation of the types for the identity and timestamp columns in
databases and files.

Previously, it was allowed to change a column type without updating the
action. Now, it is required to update the action if the type of the
identity or timestamp column is changed.

Close #432