apis/connections: simplify database rows scanning and fix `Query` method
Following the commit d63d01c, which changed the validation of identity
and timestamp columns, this commit simplifies the scanning of database
rows.

Additionally, this commit ensures that when the `Query` method is
executed, the `timestamp` column, if present, no longer necessarily
needs to be of type `DateTime`, in contrast to the `Records` method.

Fix #435