admin: align the status dots in the connections map
Currently, in the connections map, the status dots shown in the blocks
of the interlinked connections (for example, the status dot of a storage
connection linked to a file connection) are misaligned compared to the
status dots of standalone connections.

This commit resolves the issue by horizontally aligning all the status
dots on the right side of their respective block.