apis/datastore: simplify `Records` method requiring schema and toSelect
This commit simplifies the use of the `Records` method in data
warehouses by ensuring that the `schema` and `toSelect` parameters are
always required. As a result, there are no special cases where the
schema is invalid or where `toSelect` is nil.

Consequently, the commit also modifies the call and behavior of various
methods that subsequently invoke the `Records` method.