server: show absolute path of file 'panics.log' in case of panic
Since the 'panics.log' file can be located in different directories
depending on how Chichi was launched (whether as an executable from the
command line, through tests as an external process, or through tests
within the same process), the location of the "panics.log" file changes,
making it less intuitive where to look for this file in case of a panic.

This commit shows in the logged message the absolute path of that file
in case of panic.