apis/datastore: don't return the schema when reading records
Since the schema must always be passed obligatorily to the `Records`
method (see the commit c1b412732c8ece8b57dc77acc27289cbf3f1a498) - and
consequently to the `Users` and `Events` methods - there is no longer
any reason for these methods to return the schema.

Furthermore, it is not clear which properties exactly should be found
within the returned schema (for example, whether it is correct for
properties such as ID, Order, etc., to be returned).

Thus, this commit changes these methods so the schema is no longer
returned.