connector/types: disable method `Unflatten` has workaround for #448
As long as we haven't made a decision regarding whether to maintain
support for both flat and non-flat properties (see issue #289), issue #448
is preventing reads from the data warehouse that refer to
sub-properties, thus preventing the exportation to the files implemented
in PR #446 and app and db exports that refers to sub-properties in their
mappings.

This commit renders the `Unflatten` call ineffective, causing flat
information to be sent outside the APIs and, consequently, to be read
again from the APIs and to be passed to the data warehouses to build the
correct query.

Refer to issue #448 for more details.