apis: store the file export schema in the action
Currently, when exporting to a file, the schema of the exported data in
the file is read each time from the data warehouse.

This commit modifies that behavior by ensuring that the schema of users
is stored within the output schema of the action. This is done to
prevent the file structure from changing unexpectedly without the manual
confirmation of the user (who must save the action again in order to
change the exported file structure).