apis,connectors: implement iterator for writing records
This change introduces the `connectors.Writer` interface for writing
records using an iterator. This change is applied in the new
`(*Action).exportUsers` method, enabling the export of users to apps,
databases, or file connectors.

Additionally, modify the `connector.RecordReader` interface by adding a
new `Ack` method for acknowledging processed records. A new return
parameter, the GID of the user, is also added to the `Record` method.

Also, it adds the `(*Transformer).Properties` method to the
`transformers` package, allowing the reading of all the properties of a
mapping transformation.

Finally, removes an obsolete validation on the internal matching
property, since now is the data warehouse that performs the comformity
check; moreover, adds, if necessary, the internal matching property to
the field `Properties` of the query used to retrieve users from the data
warehouse.

Resolve #412

Co-authored-by: Gianluca Mondini <mondini@open2b.com>