all: implement members invitations
This commit introduces invitations for members.

It is no longer possible to manually create a member. Instead, an invitation
email must be sent, which must be accepted by the member, who will then proceed
to sign up by entering their name and password.

In the list of members, badges are displayed that mark the members who have an
invitation pending. An additional badge is shown when the invitation has
expired, namely after three days from sending.

It is also possible to send a new invitation to an email for which there is
already a pending invitation: in this way, a new invitation will be created,
making the old one invalid.

Close #452