apis, connector/types: remove the concept of "flat"
This commit removes the concept of "flat" from 'connector/types' and the
APIs, ensuring that all objects are flat. The responsibility of
supporting both "flat" and non-"flat" objects (e.g., PostgreSQL
composite types) is now delegated to individual data warehouse drivers.

It should be noted that, although every object is now considered "flat",
the "Property.Flat" field cannot simply be assumed to always be 'true'.
This is because, in addition to indicating whether an object was "flat",
it implicitly asserted that the property had the kind Object. Therefore,
as seen in the commit, checks on "Property.Flat" may have been replaced
with checks on the property's kind for this reason.

Close #289.
Close #393.
Resolve #448.