apis,connectors/postgres: remove support for composite types in Postgres
Support for composite types in PostgreSQL is implemented in the data
warehouse driver and connector without guarantees that is complete and
correct, and furthermore, it is not supported at all by the identity
resolution procedure (see #353).

For these reasons, aiming to clean up the implementation and favoring
object representation through underscores (those that were the 'flat'
objects), support for composite types is removed from both the driver
and the PostgreSQL connector.