admin: fix number conversion in connector settings
Currently, if a `number` type input is modified in the settings page of
a connector, the server returns an error.

This error is due to the fact that the UI, after a change in the input,
updates the settings with the raw value taken from the input, which is
of `String` type. Indeed, if no change is made, the original `Number`
type value sent by the server will remain in the settings, and therefore
the error will not occur.

This commit fixes the problem by performing a conversion to `Number`
before updating the settings, when the input in question is a `number`
type input.