apis/datastore: specify properties instead of columns in `Merge` method
For #465.

Note that the values are always passed as rows. This is modified
separately, as this change only involves properties definition for the
`Merge` method.

After this change, 'eventsMergeTable' contains the same values as
'events.Schema' (with the exception of 'gid', discussed in issue #476).
We could consider referring to the same data structure, but this would
open the issue of cyclic imports. Therefore, in this modification, it
was preferred to copy the properties (indicating a reference in a
comment) and postpone to a later change the possibility of sharing the
data structure that defines the properties of the events.