apis/datastore: convert snake_case columns to camelCase properties
This change transforms column names into property names, altering the
case from snake_case to camelCase and vice versa. Additionally, it
modifies the syntax in hidden columns by moving the underscore from a
prefix to a suffix.

Now, if a column has an underscore as a prefix, the corresponding
top-level property will have the first letter capitalized. This
commit also changes the properties of the Dummy connector from
snake_case to camelCase.

Closes #406