apis/datastore/warehouses/postgresql: fix use of ID and OrderBy props
Currently, the ID and OrderBy properties are inserted into the query
that returns Records without being transformed into columns. This means
that if the name of either of these two properties does not match the
column name, the query referencing that column will behave invalidly.

This commit corrects this behavior by using the column instead of the
property where necessary.