admin: fix disabled `Save` when editing destination files without sheets
Currently, when editing an action of a file destination, the `Save`
button is disabled.

This behavior is essential for files that have sheets, as it's necessary
to validate that the file in question actually exists to retrieve its
sheets before confirming the changes.

However, the disabling of the button also occurs for connections that do
not have sheets, making it impossible to save them after modification.

This commit fixes the issue by ensuring that the disabling only occurs
for destinations that have sheets.