admin: change sheets behavior inside actions of file destinations
Currently, in file destinations that support sheets, it's necessary to
fetch the sheets after entering the path, and then select the desired
sheet from a dropdown.

This commit enables manual entry of the sheet name for file
destinations, allowing users to directly write the name of the sheet
they want to export to.

Close #430