apis/datastore/warehouses: add conformity checks when writing
Currently, the schema conformity check is only done during reading from
the data warehouse and not during writing.

This commit makes write operations on the data warehouse check the
conformity of the provided schema with the one currently present in the
database. All of this, assuming that the schema remains unchanged.

Resolve #486.