apis: allow meta-properties in output schema for destination actions
At the moment, it is not possible to add a destination action on a
connection of type file from the UI because its output schema would
contain the 'Id' property, which, being a meta-property, cannot exist in
such a schema.

This commit ensures that the creation/update of actions containing
meta-properties in the output schema is prevented only if they write to
the data warehouse; otherwise, properties whose name starts with an
uppercase letter must be accepted.

Furthermore, this commit makes some changes to emphasize, consistently
with what is stated in the documentation in 'doc', that
"meta-properties" are a concept related to data warehouses. For example,
if an app exposes a property starting with an uppercase letter, it is
not considered a meta-property.