apis/datastore/warehouses/postgresql: fix the identity resolution
Currently, when importing two users with the same external ID from two
different connections/actions, the identity resolution implemented in
PostgreSQL mistakenly combines them. This causes bug #493.

This commit corrects this behavior and also adds a more comprehensive
test on the identity resolution process and the creation of users and
user identities.

Fix #493.