apis, database: make the identity resolution work on connections, not actions
So far, identity resolution has operated on identities identified through an
action. This was implemented this way because identifiers were at the action
level, and therefore, identities had to be imported at the action level to be
properly managed.

We then moved identifiers to the workspace level, and since then, identities
have remained associated with actions.

Now that there are no longer any logical or implementation issues, and indeed,
the logic with which an imported identity is associated with a connection is
clear, this commit ensures that identity resolution operates on connections
rather than actions.