apis/datastore/warehouses/postgresql: fix resolution of events identity
In the case of an identity originating from a non-anonymous event, this
identity should be merged with other identities of the same connection
that (1) share the same user ID or (2) are anonymous and have, among
various anonymous IDs, the same anonymous ID.

Currently, however, a non-anonymous identity is merged with other
non-anonymous identities with a different user ID just because they
share an anonymous ID, overwriting them. This is not the expected
behavior.

This commit corrects this behavior.