apis/events: fix "traits" and "properties" in method `storeEvents`
Currently, both "properties" and "traits" are built from a buffer shared
among various incoming events, which is reset each time. The issue #502
depends on calling the 'Bytes' method, which instead of returning a
copy, returns a reference to the slice. Consequently, subsequent method
calls on the buffer corrupt the values of "properties" and "traits".

This commit corrects this behavior by making a copy of the slice.

Fix #502.

Co-authored-by: Marco Gazerro <gazerro@open2b.com>