apis/events,trace-events-script: add CORS support
This change adds CORS support into the collector. Alongside this, it
adjusts the content type sent by the browser when dispatching an event,
shifting from "application/json" to "text/plain." This ensures that the
request qualifies as a "simple request," preventing the browser from
triggering a CORS preflight. Simply put, the browser won't initiate an
OPTIONS request before the POST request.

Additionally, the collector now responds with an error if the request
method is not "POST" or "OPTIONS."