trace.events.script: ensure correct builds by specifying target
Previously, the 'minify' step for building 'chichi.min.js' lacked a
specified target, causing esbuild to default to 'esnext.' This resulted
in the minified script being wrapped with an arrow function incompatible
with non-ES5 browsers. The same issue occurred in the 'build:es6'
script.

This commit rectifies the issue by setting 'es5' as the target for
'minify' and 'es6' as the target for 'build:es6'.