trace-events-script: make a duplicate of the events before dispatching
The `send` method of `Sender` currently lacks a mechanism to duplicate
the event, allowing it to be altered after the `send` call has returned
but before the event is effectively dispatched.

This change ensures that the `send` method promptly serializes the event
into JSON before dispatching. Furthermore, the `sentAt` field is now
sent to the server not for each individual event but once for the entire
set of events.