apis, database: stop writing an empty GID when flushing events
Currently, when flushing events to the database, a value of "0" is
written to the "gid" column, as the actual value will be calculated
later by the identity resolution procedure.

Writing the "0" is necessary because the database scripts do not provide
a default value for the "gid" column. Since this is an implementation
issue rather than a logical one, this commit ensures that the database
SQL scripts include a default value, eliminating the need to pass a GID
of zero during event flushing to the database.