apis: standardize the use of identifiers as paths instead of names
Currently, in Chichi, there is an inconsistency in how identifiers are
treated, as in some instances, they are regarded as simple property
names, while in others, they are treated as paths referring to
properties. Refer to issue #514 for more details on this.

This commit standardizes the behavior by handling identifiers as
property paths, consistent with how they were already validated.

Resolve #514.