trace-events-script: send events with `sendBeacon` on page unloading
Firefox does not support the 'keepalive' option of the 'fetch' function.

This change, when keepalive is required and sendBeacon is supported,
uses the sendBeacon method instead of fetch.

Note that Chrome, starting from version 59, requires that the 'type'
option of the Blob to sent as the body with sendBeacon is a CORS
compatible content type.

Additionally, fix the 'onUnload' function for proper functionality.