apis: use the term "Workspace Identity Resolution" consistently
A proper term will be decided and used in the code and the
documentation.

For now, this commits deprecates the term "ResolveSyncUsers", which was
improper, and use "RunWorkspaceIdentityResolution" in its place.

The term "Workspace" differentiates this procedure from the Connection
Identity Resolution, which may be also renamed in the future.