all: formalize types that can be identifiers
Currently, in Chichi, there is no implemented control on properties that
can be identifiers or not. Therefore, it is possible to select any
property from the UI to use it as an identifier. Additionally, unclear
errors or even unexpected behaviors happen (see the issue #321).

This commit standardizes and formalizes the behavior.

Resolve #321.

Also see #524.