apis/datastore, doc: remove support for Array as identifiers
Currently, the implementation of Arrays as identifiers in the Workspace
Identity Resolution treats them like any other type, resulting in
potentially unexpected or incorrect behaviors.

This commit removes support for Arrays as identifiers so that we can
consider reintroducing them in the future along with a correct and
tested implementation.