all: improve the warehouse users UI
This commit enhances the UI for warehouse users. Now, user details are displayed
through a side drawer instead of a full-screen page.

In the drawer, it is possible to view the user's traits, associated events, and
identities. Additionally, the drawer allows for user list navigation using
specific navigation arrows, and the row of the currently viewed user will be
highlighted with a background color.

Furthermore, 'displayed properties' were added, which are special properties
mapped from the user schema. These can be optionally configured in the workspace
settings to show a distinctive user header featuring their image, first name,
last name, and an additional piece of information of choice.

Close #494