database/warehouses: fix declaration of meta properties for identities
Currently, meta properties are implicitly declared as nullable, even
though in the code they are treated as if they always have a non-null
value.

This commit aims to standardize this inconsistency, particularly because
when requesting a property using the 'Records' method and the schema
passed to that method does not exactly match the one in the data
warehouse (for example, regarding the Nullable field of a property),
the returned error is not easy to identify. In this regard, we should
consider a solution to identify this type of bug/inconsistency.