apis/datastore/warehouses: consider nullability in `CheckConformity`
Currently, the `CheckConformity` function does not take into account the
nullability of a property when performing the conformity check.

This causes unexpected behaviors, panics, or even difficult-to-debug
errors when nullability does not match, as observed in commit e9ebff23e33d212d0ae3c1a831288e76be08c263.

This commit ensures that nullability is considered when checking the
conformity of schemas.

It then modifies all tests which refer to the warehouse schema to make
them consistent and working again.

Also refer to issue #527, mentioned in the code, regarding the
nullability of Objects obtained from flat properties.