apis/datastore/warehouses/postgresql: fix Identity Resolution query
Currently, when executing tests, an error may be reported about a "null"
constrain violation on the column "_gid" of "users_identities", causing
the tests to fail.

This may happen when the Workspace Identity Resolution is executed
contemporarily the writing of users identities on the warehouse.

This commit fixes a query to avoid that this error may happen. However,
we should test how the Identity Resolution behave when executed more
than once concurrently.

This fix does not affect the issue #550.