trace-events-script: embed write key in storage key names
This commit embeds the first 7 letters of the write key into both the
keys of the storage and the key of the persisted queue. By doing so, it
allows for the utilization of more write keys on the same page, while
ensuring that a new write key does not read and write the values
associated with an old one.