trace-events-script: fix reading page 'url' and 'path' from canonical
When a '<link rel="canonical">' tag is present, the 'page.url' and
'page.path' fields of an event are now properly extracted from the
'href' attribute of the 'link' element. Previously, the code was flawed,
as it incorrectly retrieved the value of the attribute instead of
accessing the 'href' property of the element, which contains the
complete URL derived from the attribute value.