apis/datastore, database: add one property in users schema by default
Since now we initialize users / users_identities schemas without
properties (except for the meta properties), this implies that, within
Chichi, at any point where the schema needs to be handled, we may
encounter an invalid schema (because we don't allow valid schemas
without properties).

This causes panics in the code, for example when removing the
meta properties from the schema and obtaining an "empty schema".

So, this commit adds a workaround for #572, until it is discussed and
resolved.

Also, take a look at #573, which is partially related to this.