trace-events-script: fix how responses from sendBeacon are handled
Previously, when sending events with sendBeacon the request was
considered failed and consequently an error message was logged in the
console (where debug was true), and the events were discarded from the
queue.

This change, when sending events with sendBeacon, does not log the error
message and does not discard the events.