connector/types: revert character length for `JSON` type
This change reverts the commit "connector/types: allow setting character
length for the JSON type" (274d51888203494dad981a65ff9b9a7cdc22e880).

The character length on the 'JSON' type is not consistently supported
across various normalizations and conversions, and it would be complex
to implement due to changes in the internal Go type for the 'JSON' type
since the modification was made. Previously, the internal type for the
'JSON' type was solely 'json.RawMessage'.