apis/datastore/warehouses/postgresql: fix db connection in `AlterSchema`
Currently, the `AlterSchema` method directly accesses the 'db' inside
the 'warehouse,' and this is incorrect.

In the case that a method of the driver has already been called
previously, there is no issue as 'db' has already been initialized.

However, if the 'AlterSchema' method is the first to be called, a panic
occurs.

This commit corrects the method.