admin: fix error when saving transformation functions
Currently, when trying to create or save an action with a transformation
function, the error "output schema is required by the transformation" is
shown.

The issue stems from the fact that in commit 25772a4, which removes
anonymous identifiers, we removed the code that checks if an output
schema property of the action type is an anonymous identifier before
adding it to the `ActionToSet` output schema, but we left its
initialization with a zero value, instead of passing the entire output
schema of the action type.

This commit fixes the issue.

Close #551.