admin: update exception handling in the identifiers route
This commit ensures that an appropriate message is displayed in the
identifiers route when the user schema currently in use does not contain
any property that can be used as identifier.

Furthermore, the possibility that no data warehouse has been connected is
no longer handled, as the `NotConnected` error is no longer returned by
the server.

Close #524.