admin: add the copy-and-paste SDK snippet
This commit adds the Javascript SDK snippet to the settings of the
connections that support it.

Now it's possible to copy the snippet and insert it directly into the
HTML to start sending events. The snippet is already set up and requires
no further modifications, as the api key and the endpoint are already
pre-filled.