connector: introduce the `EventRequest` method for apps
Previously, app connectors were required to implement the
PreviewSendEvent and SendEvent methods to send events to the app. With
this update, they now only need to implement the EventRequest method.

The EventRequest method returns a request containing the event to be
sent to the app, which is then handled directly by the dispatcher.