admin,apis: fix panic caused by invalid property names in query results
Previously, executing a query with the database connector would trigger
a panic if the results contained a column name that couldn't be
represented as a property name. This change resolves the panic and
provides clear error messages for both invalid property names and
repeated property names.

Additionally, it addresses the handling of unprocessable errors in the
UI that may arise during query execution.

Fixes #607

Co-authored-by: Andrea Retini <retini@open2b.com>