apis: remove the `businessIDName` parameter from `App.Users`
The Business ID name is now taken and stored by the method
'connectors.App', when the 'App' type is initialized. That's because the
Business ID is a information stored in the app connection, and it's
useless and leads to confusion the passing of that parameter to the
method 'App.Users'.

This commit is a part of a series of changes that makes clearer the
infrastructure of the connectors.