apis: refactor the parameters of the `File.Records` method
This commit is a part of a series of changes that makes clearer the
infrastructure of the connectors.

This, in particular, removes parameters from the method `File.Records`
which may be read from the file action; keeping them as parameters here
it's useless and would just lead to confusion.

It also rewrite the documentation of the method.