apis/datastore: add workaround for #620
This commit adds a workaround for the issue:

	data warehouse error: ERROR: duplicate key value violates unique constraint "pg_proc_proname_args_nsp_index" (SQLSTATE 23505)

that seems to happen because multiple instances of the Workspace
Identity Resolution modify some FUNCTION / PROCEDURE concurrently (in
particular, the 'pg_proc' table), causing that bug.

For now, the issue #620 can be closed, as we have the issue #354 where
we discuss the execution of the Workspace Identity Resolution.

Resolve #620.