apis: remove Business ID from `FileStorage.Read` parameters
Since the Business ID of the connections no longer applies to files
(which have the Business ID in the action, instead), the Business ID is
not available during the preview of a file (because (1) it is not
present in the FileStorage connection and (2) the user chooses the
Business ID only after seeing the file preview).

So, this commit removes the Business ID from the 'FileStorage.Read'
method.