all: remove redundant `SourceDescription` and `DestinationDescription`
Previously, connectors for files, file storages, and databases were
required to provide source and destination descriptions. However, these
descriptions are redundant as they are the same for all connectors of
these types.

This change removes the `SourceDescription` and `DestinationDescription`
fields from the relevant types of files, file storages, and databases.

Additionally, it modifies the state load process to exclude populating
the `SourceDescription`, `DestinationDescription`, `TermForUsers`, and
`TermForGroups` fields of the Connection for files, file storages, and
databases.

Closes #640