chichi: refactor interfaces and methods implemented by app connectors
This change addresses several issues related to the interfaces and
methods utilized by app connectors.

Previously, app connectors were required to return the schema of event
types each time the `EventTypes` method was called. This change resolves
this issue by removing the schema from the `EventType` type and
modifying the `Schema` method to return the schema of event types.

Furthermore, app connectors were obligated to implement the `Resource`
method, even when the resource was not supported. This change rectifies
this by moving the `Resource` method into its own interface,
`AppResource`.

Additionally, this commit documents the `EventType` type and its fields,
relocating its declaration to `chichi/apps.go` for better organization.

Moreover, it enhances the documentation for the `AppEvents.EventRequest`
and `(*App).EventRequest` methods, improving clarity by renaming the
`eventType` parameter to `typ` and the `data` parameter to `extra`.