apis/datastore/warehouses/postgresql: fix #654
After investigating #654, it emerged that the bug is caused by a faulty
query in the Connection Identity Resolution, which, when merging two
identities, only considers the properties set in the latest arrival,
thus rendering other properties valueless.

This commit resolves the issue with inefficient but logically correct
code. However, when we revisit the optimization of PostgreSQL
(issue #627), this part will need to be rewritten from scratch.

Nevertheless, Chichi's behavior now seems to be as expected, and this
commit adds a test for it.

Fix #654.