ui: use `workspace.ServeUI` on action's file connector change
Currently in the UI, when editing a file action and changing the file
connector, an HTTP call is made to retrieve the UI of the file connector
through the API method `action.ServeUI`.

However, this method does not allow for the UI of the newly file
connector to be returned, but instead returns the UI of the file
connector with which the action had been previously saved.

This commit solves the problem by ensuring that this method is only
called when opening a previously saved action, and that, instead, when
changing the file connector, the method `workspace.ServeUI` is called.