ui, apis: review the special properties/columns of the action
This commit reviews the UI section that allows defining special
properties (unique ID, displayed ID, and updated at) in actions that
support them.

These properties have been renamed in the UI, as well as moved out of the
transfomation section. To improve usability, we've implemented automatic
filling of the displayed ID with the unique ID value, and for actions on
app connections, the displayed ID will autofill if the app connector
recommends a property name as the displayed ID.

We've introduced validation checks to ensure that only properties of the
appropriate type are displayed in the combobox lists for selecting
special properties. Further validation has been added to effectively
manage error conditions and user entries that are invalid.

Additionally, this commit has updated the handling of changes to action
settings in scenarios where they are crucial for the dynamic loading of
the action's schema. Now, when such settings are modified, a tooltip is
displayed next to the settings confirmation button, and the entire
transformation section, along with the special properties, becomes
disabled. Lastly, the settings are no longer directly exposed in the
action but can only be retrieved by calling the `action.ServeUI` method.

See #631.