chichi: remove again the `go.work.sum` file from the repository
The 'go.work.sum' file has been added again by the commit
42d2ad2bfd77885a1bbff22bcb298e1af98da5be, but the problems we had in the
past still persist: it's not clear how it should be or what it should
contain, how to change it, and how to put it in a
deterministic/normalized form.

That file seems to be undocumented in the Go documentation, so this
leads to more confusion.

So, this commit removes it again from the repository.