apis/datastore:warehouses/diffschemas: change behavior for Objects
This commit alters the behavior of schema diffs for Objects, ensuring
that it's not possible to add new properties to existing Objects.

Eventually, in the future, we will consider allowing the addition of
properties at any position, even amidst existing properties, and we will
then consider reintroducing the ability to extend existing Objects. This
would technically be feasible now, but extending existing Objects would
be limited only to those "at the bottom" of the properties list, and it
would be possible to add properties only at the end. This, clearly, is a
limited use case that is not worth supporting.

In addition to changing the implementation and tests, this commit also
modifies the documentation.