apis/connectors: fix UI values validation in `ValidateUIValues` method
A connector with UI is authorized to call `SetSettings` with `nil`
settings.

Previously, the ValidateUIValues method of `*Connectors` returned an
error in this situation. Additionally, it did not validate that the
settings are valid UTF-8 runes.

This change fixes these issues. It also makes the ValidateUIValues
method panic if the connector has no UI instead of returning an error.