apis: don't check warehouse in reading connections and action schemas
Previously, when invoking the `(*Workspace).Connection` and the
`(*Connection).ActionSchemas`, and if the workspace did not have a
data warehouse, the methods returned an unprocessable error. However,
these methods do not require the presence of a warehouse.

This change removes the unnecessary warehouse checks from these methods,
ensuring that they execute without needing a warehouse to be specified.