cmd/chichi: add package `stdconnectors`
This commit adds the 'stdconnectors' package, which contains the imports
of Chichi's connector packages.

This simplifies the scenario where one wants to compile Chichi into a
separate module while still maintaining flexibility. Refer to the
documentation for further details.

Furthermore, this commit ensures that this package is also imported by
the tests, as currently the testing mode that Chichi executes within the
same process - useful for debugging - is not functioning, as it was
broken when the Chichi executable was refactored.