apis/datastore: use the term "identity key" in PostgreSQL
The aim of this commit is to introduce a specific term for the value we
use to uniquely identify an identity (which is not something we directly
expose, it's a concept used internally only, different from the
"identity id" for how we intend it in Chichi), so that the name
"identity id" can be "freed up" to use it instead of the external ID.

It's also possible that when we review the implementation of Identity
Resolution, there might no longer be a need for this unique key to
identify the identities (see issue #687 for this), further simplifying
it.

For #639.