commit, vendor: add - again - the Go vendor to the repository
This commit attempts to re-add the Go vendor to the repository, after
commit d752a8e18a80965fc00847104924c17f308ecc0d removed it due to issues
with the JetBrains Goland editor.

Please note that this time the vendor is not added module by module, but
at the workspace level, using the 'go work vendor' command. This appears
to be the only correct method when working in a Go workspace with
version 1.22 of Go.