assets: add a `.gitignore` that ignores `node_modules`
The `assets` directory should not contain arbitrary files created by the
user, so it's reasonable and convenient to add a `.gitignore` that
excludes the `node_modules` directory.

If the user needs to exclude other files, they can still add them to the
`.gitignore` in the root of the repository or the global `.gitignore` of
their system.