chichi: fix problems with vendoring and `.gitattributes`
Due to line ending issues altered by Git on Windows from the
.gitattributes directives present in modules within the vendor
directory, this commit removes the vendor's .gitattributes files from
the repository, ignoring them, ensuring Git disregards and doesn't apply
the directives within them.

Note that this is a test commit, and the proper functionality of this
solution should be tested and verified on other platforms.