chichi: document the command `go run ./commit`
Instead of launching:

    go run commit/commit.go

this commit changes the documentation to refer to the form:

    go run ./commit

which is simpler to type when typing the command manually.

Note that the form:

    go run github.com/open2b/chichi/commit

is not documented because it would not work in directories that are not
the root directory of the repository, as the 'commit/commit.go' script
strictly checks the directory from which it has been launched.
This may be changed in a future commit, as it requires additional work.